+++ packages/core/src/devtools/getDebugTree.ts
@@ -35,7 +35,7 @@ export function getDebugTree(debugSequenceHash: SequenceHash, parentNode?: Debug
       const mapData = debugData.getSequenceDetails(debugSequenceHash);
       if (mapData) {
         node.slot = mapData.slotName;
-        node.sourceMap = mapData.sourceMap;
+        node.sourceLoc = mapData.sourceLoc;
       }
 
       const cssRule = debugData.getCSSRules().find(cssRule => {
