+++ packages/core/src/devtools/index.ts
@@ -1,5 +1,6 @@
 export { injectDevTools } from './injectDevTools';
 export { isDevToolsEnabled } from './isDevToolsEnabled';
 export { debugData } from './store';
+export { getSourceLocation } from './getSourceLocation';
 
-export type { DebugCSSRules, DebugSequence, DebugResult, DebugSourceMap } from './types';
+export type { DebugCSSRules, DebugSequence, DebugResult, DebugSourceLoc } from './types';
