+++ packages/core/src/index.ts
@@ -72,4 +72,4 @@ export type {
 } from './types';
 
 // Private exports, are used by devtools
-export type { DebugCSSRules, DebugSequence, DebugResult, DebugSourceMap } from './devtools';
+export type { DebugCSSRules, DebugSequence, DebugResult, DebugSourceLoc } from './devtools';
