+++ packages/devtools/src/sourceMap/index.ts
@@ -1,0 +1,1 @@
+export { loadOriginalSourceLoc } from './loadOriginalSourceLoc';
