+++ packages/devtools/src/utils.ts
@@ -3,13 +3,13 @@ import { SlotInfo } from './types';
 
 export function getRulesBySlots(node: DebugResult, result: SlotInfo[] = []): SlotInfo[] {
   if (node.children.length === 0 && node.slot) {
-    const { debugClassNames, rules, slot, sourceMap } = node;
+    const { debugClassNames, rules, slot, sourceLoc } = node;
 
     return [
       ...result,
       {
         slot,
-        sourceMap,
+        sourceLoc,
         rules: debugClassNames.map(({ className, overriddenBy }) => {
           return {
             cssRule: rules![className],
@@ -30,11 +30,11 @@ export function filterSlots(slots: SlotInfo[], searchTerm: string) {
     return slots;
   }
 
-  return slots.reduce<SlotInfo[]>((acc, { slot, rules, sourceMap }) => {
+  return slots.reduce<SlotInfo[]>((acc, { slot, rules, sourceLoc }) => {
     const filteredRules = rules.filter(rule => rule.cssRule.includes(searchTerm));
 
     if (filteredRules.length) {
-      return [...acc, { slot, rules: filteredRules, sourceMap }];
+      return [...acc, { slot, rules: filteredRules, sourceLoc }];
     }
 
     return acc;
